Moved user context to common folder
